requirements.txt
argon2-cffi==20.1.0
aspose-cells==21.7.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cffi==1.14.6
colorama==0.4.4
colormap==1.0.3
config2py==0.0.3
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dol==0.1.29
entrypoints==0.3
et-xmlfile==1.1.0
excel2img==1.4.0
grub==0.0.11
i2==0.0.32
importlib-resources==5.2.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
lined==0.0.30
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.1
numpy==1.21.1
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py2store==0.1.2
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.7.1
sheet==0.0.5
six==1.16.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
xlwt==1.3.0
zipp==3.5.0
